auth-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
pydantic==2.7.1
